+++ src/factory/race/Elf.java
@@ -1,7 +1,7 @@
 package factory.race;
 
 public class Elf extends Race {
-    public Elf(int strength, int intelligence, int agility, int toughness) {
+    public Elf(double strength, double intelligence, double agility, double toughness) {
         super(strength, intelligence, agility, toughness);
     }
 }
