+++ src/factory/race/Human.java
@@ -1,8 +1,7 @@
 package factory.race;
 
 public class Human extends Race {
-
-    public Human(int strength, int intelligence, int agility, int toughness) {
+    public Human(double strength, double intelligence, double agility, double toughness) {
         super(strength, intelligence, agility, toughness);
     }
 }
